# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Recetas DWES",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/recetas": {
            "get": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Listado de recetas",
                "operationId": "4769361c21c600e2df668b95a7c04694",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Crear una nueva receta",
                "operationId": "42db13bb7156194a5f5dbec70e4392eb",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Receta creada correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/recetas/{id}": {
            "get": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Ver una receta específica",
                "operationId": "1ff61f91c55f464ab8afd758c1900f9e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta obtenida correctamente"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Actualizar una receta",
                "operationId": "c71058c8cc3d6d35d5337302d638daad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {},
                "responses": {
                    "200": {
                        "description": "Receta actualizada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Receta no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Recetas"
                ],
                "summary": "Eliminar una receta",
                "operationId": "bf47e92166384fa00397bcffa91088cc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Receta eliminada"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}